# apps/api/src/swagger/api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/create-password": {
      "post": {
        "operationId": "AuthController_createPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePasswordBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "ZPI API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "RegisterBody": {
        "type": "object",
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "maxLength": 64,
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "archivedAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "createdAt",
              "archivedAt",
              "updatedAt",
              "email",
              "firstName",
              "lastName",
              "role"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "maxLength": 64,
            "type": "string"
          },
          "rememberMe": {
            "type": "boolean"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "archivedAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "createdAt",
              "archivedAt",
              "updatedAt",
              "email",
              "firstName",
              "lastName",
              "role"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
      "ForgotPasswordBody": {
        "type": "object",
        "properties": {
          "email": {
            "format": "email",
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "CreatePasswordBody": {
        "type": "object",
        "properties": {
          "password": {
            "minLength": 8,
            "maxLength": 64,
            "type": "string"
          },
          "createToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "password",
          "createToken"
        ]
      },
      "ResetPasswordBody": {
        "type": "object",
        "properties": {
          "newPassword": {
            "minLength": 8,
            "maxLength": 64,
            "type": "string"
          },
          "resetToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "resetToken"
        ]
      },
...
    }
  }
}